Changed entry point to index.tsx, enable hot reloading
