+++ package.json
@@ -66,6 +66,7 @@
     "@types/aphrodite": "^0.5.6",
     "@types/react": "^16.0.18",
     "@types/react-dom": "^16.0.2",
+    "@types/react-hot-loader": "^3.0.5",
     "@types/react-redux": "^5.0.10",
     "@types/react-router-dom": "^4.0.8",
     "@types/react-router-redux": "^5.0.8",
